# Perf: remove version from nodes